Fix values for k8s-infra (#484)
The decision to change the default settings for logging and OTLP
exporters, reportedly due to CI concerns, seems misguided. The OTLP
exporter should remain enabled by default. Users of k8s-infra should
have their data collected and sent to SigNoz or other OTLP backends.
With current defaults, users will not have any idea that their data is
being collected but not sent anywhere.

Being able to perform simple install and provide an endpoint to send
data is a great onboarding experience and we should strive to keep the
experience simple with sensible defaults.