Update resourcedetection config (#533)
- Recently, we got rid of `resourcedetection/internal` from both
`otelAgent` and `otelDeployment`. This change created data problems for
any resources that were not in the same node as the otelDeployment pod.
For example, 1. `k8s.node.name` was getting replaced with the node name
where deployment was running 2. The `k8s.pod.phase` has a `host.name`
attribute populated by the `host.name` of the otelDeployment pod. The
`otelAgent` and `otelDeploment` do very different jobs, so they
shouldn't apply the same resource detection config.
- While troubleshooting, I came across the `override: true`. I asked if
we know any reason for this but it doesn't seem we don't recall anything
about it
https://signoz-team.slack.com/archives/C02T94VFZAA/p1728970838055629. I
looked at the old config generated with `override` set to `true` and it
didn't make any sense
https://github.com/SigNoz/charts/issues/532#issuecomment-2413220233. It
is used in all the pipelines, I don't see any case where we would want
to override the incoming data.


In this change, I disabled the `override` and defined new resource
detection for deployment.