feat(signoz): add smtpVars options in queryService to use existing k8s secret (#577)
This change will allow users to reference an existing secret to inject
SMTP credentials for user invitation email feature. This will help in
not exposing the credentials in plain text e.g. in git.

User can either user additionalEnvs to put `SMTP_` env variables
directly in values.yaml or they can refer to a k8s secret and specify
which keys in secret to use for which SMTP env variable. The chart will
complain if SMTP variables are present both in smtpVars in
additionalEnvs to avoid overwrites of env.

README configuration list also updated for new variables.

This PR will solve the issue mentioned here:
https://github.com/SigNoz/charts/issues/443