Add metadata exporter to pipelines (#605)
Adding the exporter to the pipeline doesn't make it start working. It
needs to be enabled explicitly with `enabled: true`. That sounds
odd/anti-pattern but it seemed like the best way forward for me. Will
enable this internally in a phased manner and then make it enabled by
default in the upcoming release.