feat: Add annotations support to schema migrator templates
Enhances the schema migrator templates to support annotations, allowing for greater customization and flexibility in deployment configurations. One major use-case for this change is when the namespace has Istio injection enabled, and init containers require istio injection to be turned off to complete the jobs.

This change updates both async and sync migration templates to include annotations specified in the Helm values.